new graphics, health system, and others
I went back and made a bunch of changes and added features to the game.
Some music and graphics are different. I also added a health system and
a new goal to the game. It also now saves the high score to a file.
Lastly, I changed some movement properties and how the score is counted.